make pipelines cleaner (#309)
* clean pipelines

* use logger.info instead of print
