+++ wren-ai-service/src/globals.py
@@ -22,11 +22,9 @@ from src.pipelines.ask import (
 from src.pipelines.ask import (
     sql_correction_pipeline as ask_sql_correction_pipeline,
 )
-from src.pipelines.ask.components.prompts import text_to_sql_system_prompt
 from src.pipelines.ask_details import (
     generation_pipeline as ask_details_generation_pipeline,
 )
-from src.pipelines.ask_details.components.prompts import ask_details_system_prompt
 from src.pipelines.semantics import description
 from src.utils import init_providers
 from src.web.v1.services.ask import AskService
@@ -36,6 +34,8 @@ from src.web.v1.services.semantics import SemanticsService
 SEMANTIC_SERVICE = None
 ASK_SERVICE = None
 ASK_DETAILS_SERVICE = None
+SQL_EXPLANATION_SERVICE = None
+SQL_REGENERATION_SERVICE = None
 
 
 def init_globals(
@@ -43,19 +43,20 @@ def init_globals(
         [], Tuple[LLMProvider, DocumentStoreProvider]
     ] = init_providers,
 ):
-    global SEMANTIC_SERVICE, ASK_SERVICE, ASK_DETAILS_SERVICE
+    global \
+        SEMANTIC_SERVICE, \
+        ASK_SERVICE, \
+        ASK_DETAILS_SERVICE, \
+        SQL_EXPLANATION_SERVICE, \
+        SQL_REGENERATION_SERVICE
 
     llm_provider, document_store_provider = init_providers()
-    ddl_store = document_store_provider.get_store()
 
     SEMANTIC_SERVICE = SemanticsService(
         pipelines={
             "generate_description": description.Generation(
-                embedder=llm_provider.get_text_embedder(),
-                retriever=document_store_provider.get_retriever(
-                    document_store=ddl_store
-                ),
-                generator=llm_provider.get_generator(),
+                llm_provider=llm_provider,
+                document_store_provider=document_store_provider,
             ),
         }
     )
@@ -64,44 +65,35 @@ def init_globals(
         pipelines={
             "indexing": ask_indexing_pipeline.Indexing(
                 llm_provider=llm_provider,
-                store_provider=document_store_provider,
+                document_store_provider=document_store_provider,
             ),
             "query_understanding": ask_query_understanding_pipeline.QueryUnderstanding(
-                generator=llm_provider.get_generator(),
+                llm_provider=llm_provider,
             ),
             "retrieval": ask_retrieval_pipeline.Retrieval(
-                embedder=llm_provider.get_text_embedder(),
-                retriever=document_store_provider.get_retriever(
-                    document_store=ddl_store
-                ),
+                llm_provider=llm_provider,
+                document_store_provider=document_store_provider,
             ),
             "historical_question": historical_question.HistoricalQuestion(
                 llm_provider=llm_provider,
                 store_provider=document_store_provider,
             ),
             "generation": ask_generation_pipeline.Generation(
-                generator=llm_provider.get_generator(
-                    system_prompt=text_to_sql_system_prompt,
-                ),
+                llm_provider=llm_provider,
             ),
             "sql_correction": ask_sql_correction_pipeline.SQLCorrection(
-                generator=llm_provider.get_generator(
-                    system_prompt=text_to_sql_system_prompt,
-                ),
+                llm_provider=llm_provider,
             ),
             "followup_generation": ask_followup_generation_pipeline.FollowUpGeneration(
-                generator=llm_provider.get_generator(
-                    system_prompt=text_to_sql_system_prompt,
-                ),
+                llm_provider=llm_provider,
             ),
         }
     )
+
     ASK_DETAILS_SERVICE = AskDetailsService(
         pipelines={
             "generation": ask_details_generation_pipeline.Generation(
-                generator=llm_provider.get_generator(
-                    system_prompt=ask_details_system_prompt
-                )
+                llm_provider=llm_provider,
             ),
         }
     )
