+++ wren-ai-service/src/pipelines/ask/components/post_processors.py
@@ -16,31 +16,6 @@ logger = logging.getLogger("wren-ai-service")
 
 
 @component
-class QueryUnderstandingPostProcessor:
-    @component.output_types(
-        is_valid_query=bool,
-    )
-    def run(self, replies: List[str]):
-        try:
-            result = orjson.loads(replies[0])["result"]
-
-            if result == "yes":
-                return {
-                    "is_valid_query": True,
-                }
-
-            return {
-                "is_valid_query": False,
-            }
-        except Exception as e:
-            logger.error(f"Error in QueryUnderstandingPostProcessor: {e}")
-
-            return {
-                "is_valid_query": True,
-            }
-
-
-@component
 class GenerationPostProcessor:
     @component.output_types(
         valid_generation_results=List[Optional[Dict[str, Any]]],
@@ -74,10 +49,6 @@ class GenerationPostProcessor:
                 "valid_generation_results": [],
                 "invalid_generation_results": [],
             }
-
-
-def init_query_understanding_post_processor():
-    return QueryUnderstandingPostProcessor()
 
 
 def init_generation_post_processor():
