+++ wren-ai-service/tests/pipelines/test_ask_details.py
@@ -5,16 +5,30 @@ from src.web.v1.services.ask_details import (
 )
 
 
-def test_generation_pipeline():
+def test_generation_pipeline_producing_executable_sqls():
     llm_provider, _ = init_providers()
     generation_pipeline = Generation(
-        generator=llm_provider.get_generator(),
+        llm_provider=llm_provider,
     )
 
-    generation_result = generation_pipeline.run(
-        "SELECT * FROM book",
-    )
+    candidate_sql_queries = [
+        "SELECT COUNT(*) FROM book",
+        "SELECT Writer FROM book ORDER BY Writer ASC NULLS FIRST",
+        "SELECT Title FROM book ORDER BY Issues ASC NULLS FIRST",
+        'SELECT Title FROM book WHERE Writer <> "Elaine Lee"',
+        "SELECT Title, Issues FROM book",
+        "SELECT Publication_Date FROM publication ORDER BY Price DESC",
+        "SELECT DISTINCT Publisher FROM publication WHERE Price > 5000000",
+        "SELECT Publisher FROM publication ORDER BY Price DESC LIMIT 1",
+        "SELECT COUNT(DISTINCT Publication_Date) FROM publication",
+        'SELECT Price FROM publication WHERE Publisher = "Person" OR Publisher = "Wiley"',
+    ]
 
-    assert AskDetailsResultResponse.AskDetailsResponseDetails(
-        **generation_result["post_processor"]["results"]
-    )
+    for candidate_sql_query in candidate_sql_queries:
+        generation_result = generation_pipeline.run(
+            candidate_sql_query,
+        )
+
+        assert AskDetailsResultResponse.AskDetailsResponseDetails(
+            **generation_result["post_processor"]["results"]
+        )
